requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
pillow==11.2.1
django-bootstrap-datepicker-plus==5.0.4
django-simple-history==3.4.0
django-filter==23.3
djangorestframework==3.14.0
xhtml2pdf==0.2.17
whitenoise==6.5.0
python-dateutil==2.8.2
reportlab==4.0.7
openpyxl==3.1.2
